Rename container to casparcontainer
