+++ src/casparcg/casparcontainer.js
