+++ test/casparcg/casparcontainer.js
@@ -1,4 +1,4 @@
-import CasparDeviceContainer from '../../src/casparcg/container';
+import CasparDeviceContainer from '../../src/casparcg/casparcontainer';
 
 test('should create one primary CasparDevice when initilize', () => {
   const casparCGContainer = new CasparDeviceContainer('test-name', 'localhost', 5250);
